+++ app/controllers/orders_controller.rb
@@ -1,0 +1,29 @@
+class OrdersController < ApplicationController
+  before_action :authenticate_user!, only: [:create]
+
+  def create
+    @order = Order.new(order_params)
+    @order.user = current_user
+    @order.total = current_cart.total_price
+
+    if @order.save
+       current_cart.cart_items.each do |cart_item|
+       product_list = ProductList.new
+       product_list.order = @order
+       product_list.product_name = cart_item.product.title
+       product_list.proudct_price = cart_item.product.price
+       product_list.quantity = cart_item.quantity
+       product_list.save
+     end
+
+      redirect_to order_path(@order)
+    else
+      render 'carts/checkout'
+    end
+  end
+    private
+
+    def order_params
+      params.require(:order).permit(:billing_name, :billing_address, :shipping_name, :shipping_address)
+    end
+end
